Basic annotations - add

+++ 0x00-python_variable_annotations/0-add.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
 
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> float:
     """Return the addition of a and b."""
     return a + b
